chore: moved the elasticsearch files into seperate directories
