+++ bitcoin.stackexchange.com/elasticsearch-scraper/util.py
